refactor expected value calculations
Output should be the same as before refactor, except for max cap
reward for quests. The fix for using carving god on tail carves
in the cap calculation was only applied to the
print_monsters_and_rewards copy of the code. It should be less than
the previous value by the differece between min and max for tail
carve.